SmartRoom/requirements.txt
Flask
opencv-python
numpy
face-recognition
Pillow
Flask-SQLAlchemy
Werkzeug